Identify records lacking or with invalid coordinates but containing locality information

This function Identifies records whose coordinates can potentially be extracted from locality information.

bdc_coordinates_from_locality(
  data,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  locality = "locality"
)

Arguments

data

data.frame. Containing geographical coordinates and the column "locality'.

lat

character string. The column name with latitude in decimal degrees and WGS84. Default = "decimalLatitude".

lon

character string. The column with longitude in decimal degrees and WGS84. Default = "decimalLongitude".

locality

character string. The column name with locality information. Default = "locality".

Value

A data.frame containing records missing or with invalid coordinates but with potentially useful locality information is saved in Output/Check/01_coordinates_from_locality.csv.

Details

According to DarwinCore terminology, locality refers to "the specific description of the place" where an organism was recorded.

See also

Examples

if (FALSE) {
lat <- c(NA, NA, "")
lon <- c("", NA, NA)
x <- data.frame(lat = lat, lon = lon, locality =  c("Brazil", "Argentina", "Chile"))
bdc_coordinates_from_locality(data = x, lat = "lat", lon = "lon", locality = "locality")
}