Create or update the column summarizing the results of data quality tests

This function creates or updates the column ".summary" summarizing the results of data quality tests (i.e., columns starting with "."). Records that have failed in at least one test are flagged for further inspection (i.e., flagged as "FALSE") in the ".summary" column.

bdc_summary_col(data)

Arguments

data

data.frame. Containing the results of data quality tests (i.e., columns starting with ".").

Value

A data.frame containing a new or an updated column ".summary".

Details

If existing, the column ".summary" will be removed and then updated considering all test names available in the supplied database.

Examples

.missing_names <- c(TRUE, TRUE, TRUE, FALSE, FALSE)
.missing_coordinates <- c(TRUE, FALSE, FALSE, TRUE, FALSE)
x <- data.frame(.missing_names, .missing_coordinates)

bdc_summary_col(data = x)
#> 
#> bdc_summary_col:
#> Flagged 4 records.
#> One column was added to the database.
#>   .missing_names .missing_coordinates .summary
#> 1           TRUE                 TRUE     TRUE
#> 2           TRUE                FALSE    FALSE
#> 3           TRUE                FALSE    FALSE
#> 4          FALSE                 TRUE    FALSE
#> 5          FALSE                FALSE    FALSE